Remove automake as a dependency to build PHP
The Automake and aclocal were part of the previous PHP build system
where Automake created Makefile from the Makefile.in templates and
aclocal was used to produce the aclocal.m4. Need for Automake as a
dependency to build PHP has been removed in PHP 4.3.0 release and
aclocal also in that period of 4.3 releases.