Drop "Extension API compatibilty" section
This was commented in the PHP 5 version, but apparently got
uncommented in the PHP 7 version. Just drop it entirely.

A discussion of module entries doesn't belong here, and generally
this section teaches bad practices. Extension authors should be
checking PHP_VERSION_ID only, and nothing else. That is, only
the last sentence of this section has the right advice.